+++ internal/fleet/package_resource_test.go
@@ -1,0 +1,43 @@
+package fleet_test
+
+import (
+	"testing"
+
+	"github.com/elastic/terraform-provider-elasticstack/internal/acctest"
+	"github.com/elastic/terraform-provider-elasticstack/internal/versionutils"
+	"github.com/hashicorp/go-version"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+)
+
+var minVersionPackage = version.Must(version.NewVersion("8.6.0"))
+
+const packageConfig = `
+provider "elasticstack" {
+  elasticsearch {}
+  kibana {}
+}
+
+resource "elasticstack_fleet_package" "test_package" {
+  name         = "tcp"
+  version      = "1.16.0"
+  force        = true
+  skip_destroy = true
+}
+`
+
+func TestAccResourcePackage(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV5ProviderFactories: acctest.Providers,
+		Steps: []resource.TestStep{
+			{
+				SkipFunc: versionutils.CheckIfVersionIsUnsupported(minVersionPackage),
+				Config:   packageConfig,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("elasticstack_fleet_package.test_package", "name", "tcp"),
+					resource.TestCheckResourceAttr("elasticstack_fleet_package.test_package", "version", "1.16.0"),
+				),
+			},
+		},
+	})
+}